classification/pyproject.toml
[tool.poetry]
name = "ml_jobs"
version = "0.1.0"
description = ""
authors = ["John Smith <john@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
sklearn = "^0.21.0"
pandas = "^1.3.4"
fasttext = "^0.9.2"
db-lib = {path = "../db_lib"}
fastapi-sqlalchemy = "^0.2.1"
click = "^8.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"